# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/packages": {
            "get": {
                "description": "Retrieve a list of all packages",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all packages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Packages"
                            }
                        }
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "description": "Retrieve a list of all projects",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Projects"
                            }
                        }
                    }
                }
            }
        },
        "/packages/{id}": {
            "get": {
                "description": "Retrieve a package by its ID\nRetrieve a package by its ID",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "summary": "Get a package by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Projects ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Packages ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Packages"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a package by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a package",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Packages ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Packages"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a package by ID",
                "summary": "Delete a package",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "packages ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "Retrieve a project by its ID\nRetrieve a project by its ID",
                "produces": [
                    "application/json",
                    "application/json"
                ],
                "summary": "Get a project by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Projects ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Projects ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Projects"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a project by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Projects ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Projects"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a project by ID",
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Projects ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ent.Packages": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "The time the package was created",
                    "type": "string"
                },
                "description": {
                    "description": "A brief description of the package",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "link": {
                    "description": "The link to the package",
                    "type": "string"
                },
                "name": {
                    "description": "The name of the package",
                    "type": "string"
                },
                "updated_at": {
                    "description": "The time the package was last updated",
                    "type": "string"
                }
            }
        },
        "ent.Projects": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description holds the value of the \"description\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "image_url": {
                    "description": "ImageURL holds the value of the \"image_url\" field.",
                    "type": "string"
                },
                "link": {
                    "description": "Link holds the value of the \"link\" field.",
                    "type": "string"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        }
    }
}