# ent/packages.go
// Code generated by ent, DO NOT EDIT.

package ent

import (
	"fmt"
	"project-manager/ent/packages"
	"strings"
	"time"

	"entgo.io/ent"
	"entgo.io/ent/dialect/sql"
)

// Packages is the model entity for the Packages schema.
type Packages struct {
	config `json:"-"`
	// ID of the ent.
	ID int `json:"id,omitempty"`
	// The name of the package
	Name string `json:"name,omitempty"`
	// The link to the package
	Link string `json:"link,omitempty"`
	// A brief description of the package
	Description string `json:"description,omitempty"`
	// The time the package was created
	CreatedAt time.Time `json:"created_at,omitempty"`
	// The time the package was last updated
	UpdatedAt    time.Time `json:"updated_at,omitempty"`
	selectValues sql.SelectValues
}

// scanValues returns the types for scanning values from sql.Rows.
func (*Packages) scanValues(columns []string) ([]any, error) {
	values := make([]any, len(columns))
	for i := range columns {
		switch columns[i] {
		case packages.FieldID:
			values[i] = new(sql.NullInt64)
		case packages.FieldName, packages.FieldLink, packages.FieldDescription:
			values[i] = new(sql.NullString)
		case packages.FieldCreatedAt, packages.FieldUpdatedAt:
			values[i] = new(sql.NullTime)
		default:
			values[i] = new(sql.UnknownType)
		}
	}
	return values, nil
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the Packages fields.
func (pa *Packages) assignValues(columns []string, values []any) error {
	if m, n := len(values), len(columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	for i := range columns {
		switch columns[i] {
		case packages.FieldID:
			value, ok := values[i].(*sql.NullInt64)
			if !ok {
				return fmt.Errorf("unexpected type %T for field id", value)
			}
			pa.ID = int(value.Int64)
		case packages.FieldName:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field name", values[i])
			} else if value.Valid {
				pa.Name = value.String
			}
		case packages.FieldLink:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field link", values[i])
			} else if value.Valid {
				pa.Link = value.String
			}
		case packages.FieldDescription:
			if value, ok := values[i].(*sql.NullString); !ok {
				return fmt.Errorf("unexpected type %T for field description", values[i])
			} else if value.Valid {
				pa.Description = value.String
			}
		case packages.FieldCreatedAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field created_at", values[i])
			} else if value.Valid {
				pa.CreatedAt = value.Time
			}
		case packages.FieldUpdatedAt:
			if value, ok := values[i].(*sql.NullTime); !ok {
				return fmt.Errorf("unexpected type %T for field updated_at", values[i])
			} else if value.Valid {
				pa.UpdatedAt = value.Time
			}
		default:
			pa.selectValues.Set(columns[i], values[i])
		}
	}
	return nil
}

// Value returns the ent.Value that was dynamically selected and assigned to the Packages.
// This includes values selected through modifiers, order, etc.
func (pa *Packages) Value(name string) (ent.Value, error) {
	return pa.selectValues.Get(name)
}

// Update returns a builder for updating this Packages.
// Note that you need to call Packages.Unwrap() before calling this method if this Packages
// was returned from a transaction, and the transaction was committed or rolled back.
func (pa *Packages) Update() *PackagesUpdateOne {
	return NewPackagesClient(pa.config).UpdateOne(pa)
}

// Unwrap unwraps the Packages entity that was returned from a transaction after it was closed,
// so that all future queries will be executed through the driver which created the transaction.
func (pa *Packages) Unwrap() *Packages {
	_tx, ok := pa.config.driver.(*txDriver)
	if !ok {
		panic("ent: Packages is not a transactional entity")
	}
	pa.config.driver = _tx.drv
	return pa
}

// String implements the fmt.Stringer.
func (pa *Packages) String() string {
	var builder strings.Builder
	builder.WriteString("Packages(")
	builder.WriteString(fmt.Sprintf("id=%v, ", pa.ID))
	builder.WriteString("name=")
	builder.WriteString(pa.Name)
	builder.WriteString(", ")
	builder.WriteString("link=")
	builder.WriteString(pa.Link)
	builder.WriteString(", ")
	builder.WriteString("description=")
	builder.WriteString(pa.Description)
	builder.WriteString(", ")
	builder.WriteString("created_at=")
	builder.WriteString(pa.CreatedAt.Format(time.ANSIC))
	builder.WriteString(", ")
	builder.WriteString("updated_at=")
	builder.WriteString(pa.UpdatedAt.Format(time.ANSIC))
	builder.WriteByte(')')
	return builder.String()
}

// PackagesSlice is a parsable slice of Packages.
type PackagesSlice []*Packages